requirements/local.requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/local.requirements.txt requirements/local.requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
black==24.10.0
    # via -r requirements/local.requirements.in
build==1.2.2.post1
    # via pip-tools
celery==5.4.0
    # via
    #   -r requirements/local.requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==44.0.0
    # via -r requirements/local.requirements.in
dj-rest-auth==7.0.0
    # via -r requirements/local.requirements.in
django==5.1.4
    # via
    #   -r requirements/local.requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-filter
    #   django-health-check
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.3.0
    # via -r requirements/local.requirements.in
django-celery-beat==2.7.0
    # via -r requirements/local.requirements.in
django-celery-results==2.5.1
    # via -r requirements/local.requirements.in
django-cors-headers==4.6.0
    # via -r requirements/local.requirements.in
django-filter==24.3
    # via -r requirements/local.requirements.in
django-health-check==3.18.3
    # via -r requirements/local.requirements.in
django-redis==5.4.0
    # via -r requirements/local.requirements.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements/local.requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements/local.requirements.in
drf-spectacular==0.28.0
    # via -r requirements/local.requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/local.requirements.in
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements/local.requirements.in
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.48
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements/local.requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.4
    # via
    #   -r requirements/local.requirements.in
    #   pytest-django
    #   pytest-dotenv
pytest-django==4.9.0
    # via -r requirements/local.requirements.in
pytest-dotenv==0.5.2
    # via -r requirements/local.requirements.in
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
python-dotenv==1.0.1
    # via
    #   -r requirements/local.requirements.in
    #   pytest-dotenv
pyyaml==6.0.2
    # via drf-spectacular
redis==5.2.1
    # via django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r requirements/local.requirements.in
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2024.2
    # via
    #   celery
    #   django
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
